Small protobuf code reorganization
